Update Python version requirement to 3.10 or higher
The type hints used within Argus 0.0.8 is not support on Python3.9 when tested, the code 'technically' still works and most of the errors are pedantic errors around things like how the types are explicitly set i.e. def x() -> pd.DataFrame | None, or thing to do with how f-strings are used.